Be less vague.
The context of "myself" becomes unclear when the security policy is
picked up on by RSS feeds, aggregators, etc, and copied to places outside
the repository (where "myself" could possibly refer to either me or to
the owner/webmaster of said websites/places). Better to just write
"the package author" instead.